k8s: watch from latest
In a high volume k8s cluster, resource versions go stale quickly enough
that watching from a resource version returned by a pod list operation
often fails with a "too old resource version" error.

Omitting the resource version parameter from the watch call causes
k8s to watch from the latest resource version. This has been tested
empirically using our GKE clusters but the docs are ambiguous on this
point.

Styx might miss events when the watcher is re-established, but that is
already the case.

In the high-volume case it does not seem likely that there is a way to
not miss events using only the watcher, thus polling acts as a fallback.

For the low-volume case we could keep storing the last observed
resource version etc in an attempt to not miss any events. However, I
did not think that useful enough to warrant the complexity, given that
the polling mechanism will likely observe terminated pods before they
are GC'd in this case.

Note: Even omitting the resource version parameter from the watch call
has been observed to immediately fail with a "too old resource version"
error on occasion.